feat(arc-1316): fix isPending and cleanup
